requirements.txt
numpy>=1.6
llvmpy
nose
nose_parameterized
